# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Bayyina CRM API
  description: |
    REST API for Bayyina CRM — an education center management system.

    ## Authentication
    All protected endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <token>
    ```
    Tokens are obtained via `POST /api/auth/login` or `POST /api/auth/verify-otp`.
    Tokens expire after **30 days**.

    ## Roles
    | Role      | Description                                      |
    |-----------|--------------------------------------------------|
    | `admin`   | Full access to all resources                     |
    | `teacher` | Read access to students; manage groups/payments  |
    | `student` | Read-only access to own data                     |

    ## Error Format
    All errors follow this structure:
    ```json
    { "code": "errorCode", "message": "Human readable message" }
    ```

  version: 1.0.0
  contact:
    name: Mardonbek Khamidov
    email: mardonbek@bayyina.uz
servers:
  - url: http://localhost:4000/api
    description: Local development
  - url: https://api.bayyina.uz/api
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token issued on login or OTP verification. Valid for 30 days.

  # ─────────────────────────────────────────────
  # Reusable schemas
  # ─────────────────────────────────────────────
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "invalidCredentials"
        message:
          type: string
          example: "Telefon raqam yoki parol noto'g'ri"

    # ── Record ──────────────────────────────────
...
  parameters:
...
    leadId:
      name: id
      in: path
      required: true
      description: MongoDB ObjectId of the lead
      schema:
        type: string
        example: "664f1a2b3c4d5e6f7a8b9c50"
    leadUniqueLink:
      name: uniqueLink
      in: path
      required: true
      description: The lead's auto-generated 12-char hex slug
      schema:
        type: string
        example: "a1b2c3d4e5f6"
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Page number (1-based)
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Number of results per page (default 20, max 100)
...
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            tokenMissing:
              summary: No token provided
              value: { code: "tokenMissing", message: "Token mavjud emas" }
            tokenExpired:
              summary: Token has expired
              value: { code: "tokenExpired", message: "Token muddati tugagan" }
            invalidToken:
              summary: Token is malformed
              value: { code: "invalidToken", message: "Yaroqsiz token kiritildi" }
    Forbidden:
      description: Authenticated but insufficient role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "forbidden"
            message: "Ruxsat berilmadi"
...
    TooManyRequests:
      description: Rate limit exceeded (120 req / 3 min)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: "tooManyRequests"
            message: "Juda ko'p so'rovlar. Iltimos, 3 daqiqadan keyin urinib ko'ring"
...
paths:
...
  /leads/search:
    get:
      tags: [Leads]
      summary: Search leads
      description: |
        Searches leads across multiple fields and returns paginated results.
        Results are populated with `source`, `courseType`, `rejectionReason`, and `createdBy`.

        **Free-text search (`q`)** performs a case-insensitive regex match against:
        - `firstName` — lead's first name
        - `telegramId` — Telegram user ID string
        - `interest` — stated interest / desired course
        - `notes` — internal notes
        - `level` — self-reported proficiency level
        - `phone` — matched numerically (only when `q` is a valid number)
        - `age` — matched numerically (only when `q` is a valid number)

        **Field filters** can be used alongside or instead of `q`:
        | Parameter | Type | Behaviour |
        |-----------|------|-----------|
        | `status` | string | Exact match (`new` / `contacted` / `interested` / `scheduled` / `converted` / `rejected`) |
        | `source` | ObjectId | Exact match by LeadSource `_id` |
        | `paymentStatus` | string | Exact match (`unpaid` / `partial` / `paid`) |
        | `gender` | string | Exact match (`male` / `female`) |
        | `courseType` | ObjectId | Exact match by CourseType `_id` |
        | `level` | string | Partial, case-insensitive match |
        | `age` | integer | Exact age match |
        | `minAge` | integer | Age ≥ value |
        | `maxAge` | integer | Age ≤ value |

        When both `q` and field filters are provided, **both** conditions must be satisfied (AND logic).

        **Access:** Admin · Teacher.
      operationId: searchLeads
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: false
          description: |
            Free-text search term. Case-insensitive regex matched against firstName, telegramId,
            interest, notes, level, and numerically against phone and age.
          schema:
            type: string
            example: "Matematik"
        - name: status
          in: query
          required: false
          description: Filter by lead status
          schema:
            type: string
            enum: [new, contacted, interested, scheduled, converted, rejected]
        - name: source
          in: query
          required: false
          description: Filter by LeadSource ObjectId
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c11"
        - name: paymentStatus
          in: query
          required: false
          description: Filter by payment status
          schema:
            type: string
            enum: [unpaid, partial, paid]
        - name: gender
          in: query
          required: false
          description: Filter by gender
          schema:
            type: string
            enum: [male, female]
        - name: courseType
          in: query
          required: false
          description: Filter by CourseType ObjectId
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c10"
        - name: level
          in: query
          required: false
          description: Filter by self-reported level (partial match)
          schema:
            type: string
            example: "A2"
        - name: age
          in: query
          required: false
          description: Filter by exact age
          schema:
            type: integer
            example: 22
        - name: minAge
          in: query
          required: false
          description: Minimum age (inclusive)
          schema:
            type: integer
            example: 18
        - name: maxAge
          in: query
          required: false
          description: Maximum age (inclusive)
          schema:
            type: integer
            example: 35
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
          description: Page number (1-based)
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
          description: Results per page (max 100)
      responses:
        "200":
          description: Matching leads returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
                  total:
                    type: integer
                    example: 5
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 20
                  totalPages:
                    type: integer
                    example: 1
                  hasNextPage:
                    type: boolean
                    example: false
                  hasPrevPage:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: "leadsFound"
                  message:
                    type: string
                    example: "Lidlar ro'yxati olindi"
              examples:
                byNameAndStatus:
                  summary: New leads interested in maths
                  value:
                    leads:
                      - _id: "664f1a2b3c4d5e6f7a8b9b01"
                        firstName: "Sardor"
                        phone: 998911234567
                        gender: "male"
                        age: 22
                        interest: "Matematika"
                        source:
                          _id: "664f1a2b3c4d5e6f7a8b9c11"
                          name: "Instagram"
                          slug: "instagram"
                        status: "new"
                        paymentStatus: "unpaid"
                        level: "A2"
                        courseType:
                          _id: "664f1a2b3c4d5e6f7a8b9c10"
                          name: "Matematika Olympiad"
                          type: "Math"
                          direction: "Advanced"
                        createdBy:
                          _id: "664f1a2b3c4d5e6f7a8b9c0d"
                          firstName: "Alisher"
                          lastName: "Nazarov"
                    total: 5
                    page: 1
                    limit: 20
                    totalPages: 1
                    hasNextPage: false
                    hasPrevPage: false
                    code: "leadsFound"
                    message: "Lidlar ro'yxati olindi"
                byAgeRange:
                  summary: Leads aged 18–25
                  value:
                    leads:
                      - _id: "664f1a2b3c4d5e6f7a8b9b02"
                        firstName: "Dilnoza"
                        phone: 998907654321
                        gender: "female"
                        age: 20
                        status: "contacted"
                        paymentStatus: "unpaid"
                        source: "telegram"
                    total: 3
                    page: 1
                    limit: 20
                    totalPages: 1
                    hasNextPage: false
                    hasPrevPage: false
                    code: "leadsFound"
                    message: "Lidlar ro'yxati olindi"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /leads:
    get:
      tags: [Leads]
      summary: List leads
      description: |
        Returns a paginated, filterable list of leads populated with `source`, `courseType`, `rejectionReason`, and `createdBy`.

        **Access:** Admin, Teacher.

        Sorted by `createdAt` descending (newest first).

        The `search` query is a case-insensitive substring match against `firstName`.
        If the search term parses as a number, it is also matched against `phone` exactly.
      operationId: getLeads
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/pageParam"
        - $ref: "#/components/parameters/limitParam"
        - name: status
          in: query
          description: Filter by follow-up status
          schema:
            type: string
            enum: [new, contacted, interested, scheduled, converted, rejected]
            example: "new"
        - name: source
          in: query
          description: Filter by LeadSource ObjectId
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c11"
        - name: courseType
          in: query
          description: Filter by CourseType ObjectId
          schema:
            type: string
            example: "664f1a2b3c4d5e6f7a8b9c10"
        - name: paymentStatus
          in: query
          description: Filter by deposit/payment status
          schema:
            type: string
            enum: [unpaid, partial, paid]
            example: "unpaid"
        - name: search
          in: query
          description: Search by `firstName` (substring, case-insensitive) or exact `phone`
          schema:
            type: string
            example: "ali"
      responses:
        "200":
          description: Leads returned
          content:
            application/json:
              schema:
                type: object
                properties:
                  leads:
                    type: array
                    items:
                      $ref: "#/components/schemas/Lead"
                  total:
                    type: integer
                    example: 37
                  page:
                    type: integer
                    example: 1
                  limit:
                    type: integer
                    example: 20
                  totalPages:
                    type: integer
                    example: 2
                  hasNextPage:
                    type: boolean
                    example: true
                  hasPrevPage:
                    type: boolean
                    example: false
                  code:
                    type: string
                    example: "leadsFound"
                  message:
                    type: string
                    example: "Murojaatlar ro'yxati olindi"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    post:
      tags: [Leads]
      summary: Create a lead
      description: |
        Creates a new lead record.

        **Access:** Admin, Teacher.

        The server:
        - Generates a unique `uniqueLink` (12-char hex slug) used for referral-click tracking.
        - Sets `createdBy` to the caller's user ID.
        - Sets `lastActivityAt` to the creation time.

        Only `firstName` is required. `telegramId` must be unique across leads when provided.
      operationId: createLead
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
            examples:
              minimal:
                summary: Minimal lead (firstName only)
                value:
                  firstName: "Ali"
              telegramBot:
                summary: Lead coming from Telegram bot
                value:
                  firstName: "Ali"
                  phone: 998901234567
                  telegramId: "123456789"
                  source: "664f1a2b3c4d5e6f7a8b9c11"
                  interest: "English evening course"
              withCourseTypeInterest:
                summary: Lead with a target course type + scheduled call
                value:
                  firstName: "Malika"
                  phone: 998977776655
                  source: "664f1a2b3c4d5e6f7a8b9c11"
                  courseType: "664f1a2b3c4d5e6f7a8b9c10"
                  level: "A1"
                  status: "scheduled"
                  scheduledAt: "2026-04-20T14:00:00.000Z"
                  notes: "Onasi qo'ng'iroq qildi"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                type: object
                properties:
                  lead:
                    $ref: "#/components/schemas/Lead"
                  code:
                    type: string
                    example: "leadCreated"
                  message:
                    type: string
                    example: "Yangi murojaat yaratildi"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                missingFirstName:
                  value: { code: "missingField", message: "Ism kiritilishi shart" }
                invalidStatus:
                  value: { code: "invalidLeadStatus", message: "Murojaat holati noto'g'ri" }
                invalidPaymentStatus:
                  value: { code: "invalidField", message: "To'lov holati noto'g'ri" }
                invalidGender:
                  value: { code: "invalidField", message: "Jins noto'g'ri" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /leads/track/{uniqueLink}:
    get:
      tags: [Leads]
      summary: Track a referral-link click (public)
      description: |
        **Public endpoint — no authentication required.**

        Records that the lead's unique referral link was hit. On first call, sets `linkClickedAt`
        to the current time; every call bumps `lastActivityAt`.

        Intended for use from a Telegram bot, ad redirect, or landing page — anywhere the lead
        physically clicks their unique URL. The slug is long and random, so guessing another lead's
        link is not practical.

        Returns `204 No Content` on success (no response body).
      operationId: trackLeadClick
      parameters:
        - $ref: "#/components/parameters/leadUniqueLink"
      responses:
        "204":
          description: Click recorded (no body)
        "404":
          description: Lead not found for this uniqueLink
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "leadNotFound"
                message: "Murojaat topilmadi"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /leads/{id}:
    get:
      tags: [Leads]
      summary: Get a lead by ID
      description: |
        Returns a single lead populated with `source`, `courseType`, `rejectionReason`, and `createdBy`.

        **Access:** Admin, Teacher.
      operationId: getLead
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/leadId"
      responses:
        "200":
          description: Lead found
          content:
            application/json:
              schema:
                type: object
                properties:
                  lead:
                    $ref: "#/components/schemas/Lead"
                  code:
                    type: string
                    example: "leadsFound"
                  message:
                    type: string
                    example: "Murojaatlar ro'yxati olindi"
        "400":
          description: Invalid ObjectId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "invalidId"
                message: "Id noto'g'ri kiritildi"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "leadNotFound"
                message: "Murojaat topilmadi"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    put:
      tags: [Leads]
      summary: Update a lead
      description: |
        Partial update — only provided fields are changed. `lastActivityAt` is auto-bumped.

        **Access:** Admin, Teacher.

        **Typical flows:**
        - `{ "status": "contacted", "notes": "qo'ng'iroq qildim" }` — log a follow-up call.
        - `{ "status": "scheduled", "scheduledAt": "2026-04-20T14:00:00Z" }` — book a meeting.
        - `{ "status": "rejected", "rejectionReason": "<ObjectId>" }` — close the lead.
        - `{ "status": "converted" }` — mark as won; then create the `Enrollment` via `POST /enrollments`.
      operationId: updateLead
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/leadId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdateInput"
            examples:
              markContacted:
                summary: Log first contact
                value:
                  status: "contacted"
                  notes: "Birinchi marta qo'ng'iroq qildim"
              scheduleMeeting:
                summary: Book a scheduled call
                value:
                  status: "scheduled"
                  scheduledAt: "2026-04-20T14:00:00.000Z"
              rejectLead:
                summary: Close as rejected
                value:
                  status: "rejected"
                  rejectionReason: "664f1a2b3c4d5e6f7a8b9c12"
              convertLead:
                summary: Flip to converted (then create Enrollment separately)
                value:
                  status: "converted"
                  paymentStatus: "paid"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  lead:
                    $ref: "#/components/schemas/Lead"
                  code:
                    type: string
                    example: "leadUpdated"
                  message:
                    type: string
                    example: "Murojaat yangilandi"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                invalidStatus:
                  value: { code: "invalidLeadStatus", message: "Murojaat holati noto'g'ri" }
                invalidSource:
                  value: { code: "invalidField", message: "Manba noto'g'ri" }
                invalidPaymentStatus:
                  value: { code: "invalidField", message: "To'lov holati noto'g'ri" }
                invalidGender:
                  value: { code: "invalidField", message: "Jins noto'g'ri" }
                invalidId:
                  value: { code: "invalidId", message: "Id noto'g'ri kiritildi" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "leadNotFound"
                message: "Murojaat topilmadi"
        "429":
          $ref: "#/components/responses/TooManyRequests"

    delete:
      tags: [Leads]
      summary: Delete a lead
      description: |
        Permanently deletes a lead record.

        **Access:** Admin only.

        > Prefer flipping `status` to `rejected` or `converted` instead of deleting —
        > deletion loses source-attribution history.
      operationId: deleteLead
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/leadId"
      responses:
        "200":
          description: Lead deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
              example:
                code: "leadDeleted"
                message: "Murojaat o'chirildi"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: "leadNotFound"
                message: "Murojaat topilmadi"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # ══════════════════════════════════════════
  # NOTIFICATIONS
  # ══════════════════════════════════════════
...